refactor: address CodeScene findings for WhatsApp inbound attachments
- Extract _has_downloadable_attachment helper in messaging_service to
  collapse the 3-branch conditional in TurnIO/MetaCloud get_inbound_media.
- Share message envelope builders in meta_cloud_api_messages and turnio_messages
  test fixtures to remove duplicated structures.
- Split test_whatsapp_integration into focused modules
  (test_whatsapp_image_attachments, test_whatsapp_document_attachments)
  and move shared helpers to _whatsapp_attachment_helpers, keeping each
  module under CodeScene's LoC/function-count thresholds.
- Move shared WhatsApp channel fixtures to conftest.
