+++ apps/channels/tests/message_examples/meta_cloud_api_messages.py
@@ -15,7 +15,7 @@ def _wrap_in_webhook_payload(value, phone_number_id="12345"):
     }
 
 
-def text_message_value(phone_number_id="12345"):
+def _message_value(*, phone_number_id: str, message_id: str, message_type: str, payload: dict) -> dict:
     return {
         "messaging_product": "whatsapp",
         "metadata": {
@@ -26,13 +26,22 @@ def text_message_value(phone_number_id="12345"):
         "messages": [
             {
                 "from": "27456897512",
-                "id": "wamid.abc123",
+                "id": message_id,
                 "timestamp": "1706709716",
-                "text": {"body": "Hello"},
-                "type": "text",
+                "type": message_type,
+                **({message_type: payload} if payload else {}),
             }
         ],
     }
+
+
+def text_message_value(phone_number_id="12345"):
+    return _message_value(
+        phone_number_id=phone_number_id,
+        message_id="wamid.abc123",
+        message_type="text",
+        payload={"body": "Hello"},
+    )
 
 
 def text_message(phone_number_id="12345"):
@@ -40,27 +49,16 @@ def text_message(phone_number_id="12345"):
 
 
 def audio_message_value(phone_number_id="12345"):
-    return {
-        "messaging_product": "whatsapp",
-        "metadata": {
-            "display_phone_number": "+15551234567",
-            "phone_number_id": phone_number_id,
+    return _message_value(
+        phone_number_id=phone_number_id,
+        message_id="wamid.abc456",
+        message_type="audio",
+        payload={
+            "mime_type": "audio/ogg; codecs=opus",
+            "sha256": "abc123",
+            "id": "1215194677037265",
         },
-        "contacts": [{"profile": {"name": "User"}, "wa_id": "27456897512"}],
-        "messages": [
-            {
-                "from": "27456897512",
-                "id": "wamid.abc456",
-                "timestamp": "1706709716",
-                "type": "audio",
-                "audio": {
-                    "mime_type": "audio/ogg; codecs=opus",
-                    "sha256": "abc123",
-                    "id": "1215194677037265",
-                },
-            }
-        ],
-    }
+    )
 
 
 def audio_message(phone_number_id="12345"):
@@ -68,29 +66,18 @@ def audio_message(phone_number_id="12345"):
 
 
 def image_message_value(phone_number_id="12345", caption="Check this out"):
-    return {
-        "messaging_product": "whatsapp",
-        "metadata": {
-            "display_phone_number": "+15551234567",
-            "phone_number_id": phone_number_id,
+    return _message_value(
+        phone_number_id=phone_number_id,
+        message_id="wamid.img001",
+        message_type="image",
+        payload={
+            "id": "image-media-id-456",
+            "url": "https://cdn.meta.example.com/image-media-id-456",
+            "mime_type": "image/jpeg",
+            "sha256": "def456",
+            "caption": caption,
         },
-        "contacts": [{"profile": {"name": "User"}, "wa_id": "27456897512"}],
-        "messages": [
-            {
-                "from": "27456897512",
-                "id": "wamid.img001",
-                "timestamp": "1706709716",
-                "type": "image",
-                "image": {
-                    "id": "image-media-id-456",
-                    "url": "https://cdn.meta.example.com/image-media-id-456",
-                    "mime_type": "image/jpeg",
-                    "sha256": "def456",
-                    "caption": caption,
-                },
-            }
-        ],
-    }
+    )
 
 
 def image_message_value_no_caption(phone_number_id="12345"):
@@ -107,30 +94,19 @@ def document_message_value(
     filename="invoice.pdf",
     mime_type="application/pdf",
 ):
-    return {
-        "messaging_product": "whatsapp",
-        "metadata": {
-            "display_phone_number": "+15551234567",
-            "phone_number_id": phone_number_id,
+    return _message_value(
+        phone_number_id=phone_number_id,
+        message_id="wamid.doc001",
+        message_type="document",
+        payload={
+            "id": "document-media-id-789",
+            "url": "https://cdn.meta.example.com/document-media-id-789",
+            "filename": filename,
+            "mime_type": mime_type,
+            "sha256": "doc456",
+            "caption": caption,
         },
-        "contacts": [{"profile": {"name": "User"}, "wa_id": "27456897512"}],
-        "messages": [
-            {
-                "from": "27456897512",
-                "id": "wamid.doc001",
-                "timestamp": "1706709716",
-                "type": "document",
-                "document": {
-                    "id": "document-media-id-789",
-                    "url": "https://cdn.meta.example.com/document-media-id-789",
-                    "filename": filename,
-                    "mime_type": mime_type,
-                    "sha256": "doc456",
-                    "caption": caption,
-                },
-            }
-        ],
-    }
+    )
 
 
 def document_message_value_no_caption(phone_number_id="12345"):
