+++ apps/channels/tests/message_examples/turnio_messages.py
@@ -1,36 +1,59 @@
+def _vnd(
+    *,
+    author_name="User",
+    chat_uuid,
+    contact_uuid,
+    permalink_uuid,
+    message_uuid,
+    inserted_at,
+    chat_updated_at,
+    unread_count,
+    chat_inserted_at="2024-01-25T09:02:46.684610Z",
+):
+    return {
+        "v1": {
+            "author": {"id": "27456897512", "name": author_name, "type": "OWNER"},
+            "card_uuid": "None",
+            "chat": {
+                "assigned_to": "None",
+                "contact_uuid": contact_uuid,
+                "inserted_at": chat_inserted_at,
+                "owner": "+27456897512",
+                "permalink": f"https://whatsapp.turn.io/app/c/{permalink_uuid}",
+                "state": "OPEN",
+                "state_reason": "Re-opened by inbound message.",
+                "unread_count": unread_count,
+                "updated_at": chat_updated_at,
+                "uuid": chat_uuid,
+            },
+            "direction": "inbound",
+            "faq_uuid": "None",
+            "in_reply_to": "None",
+            "inserted_at": inserted_at,
+            "labels": [],
+            "last_status": "None",
+            "last_status_timestamp": "None",
+            "on_fallback_channel": False,
+            "rendered_content": "None",
+            "uuid": message_uuid,
+        }
+    }
+
+
 def text_message():
     return {
         "contacts": [{"profile": {"name": "User"}, "wa_id": "27456897512"}],
         "messages": [
             {
-                "_vnd": {
-                    "v1": {
-                        "author": {"id": "27456897512", "name": "User", "type": "OWNER"},
-                        "card_uuid": "None",
-                        "chat": {
-                            "assigned_to": "None",
-                            "contact_uuid": "08a64841-32123-4111-b91f-4ff36d676c1c",
-                            "inserted_at": "2024-01-25T09:02:46.684610Z",
-                            "owner": "+27456897512",
-                            "permalink": "https://whatsapp.turn.io/app/c/08a64841-32123-4111-b91f-4ff36d676c1c",
-                            "state": "OPEN",
-                            "state_reason": "Re-opened by inbound message.",
-                            "unread_count": 1,
-                            "updated_at": "2024-01-31T14:01:56.486313Z",
-                            "uuid": "08a64841-32123-4111-b91f-4ff36d676c1c",
-                        },
-                        "direction": "inbound",
-                        "faq_uuid": "None",
-                        "in_reply_to": "None",
-                        "inserted_at": "2024-01-31T14:01:56.467839Z",
-                        "labels": [],
-                        "last_status": "None",
-                        "last_status_timestamp": "None",
-                        "on_fallback_channel": False,
-                        "rendered_content": "None",
-                        "uuid": "08a64841-32123-4111-b91f-4ff36d676c1c",
-                    }
-                },
+                "_vnd": _vnd(
+                    chat_uuid="08a64841-32123-4111-b91f-4ff36d676c1c",
+                    contact_uuid="08a64841-32123-4111-b91f-4ff36d676c1c",
+                    permalink_uuid="08a64841-32123-4111-b91f-4ff36d676c1c",
+                    message_uuid="08a64841-32123-4111-b91f-4ff36d676c1c",
+                    inserted_at="2024-01-31T14:01:56.467839Z",
+                    chat_updated_at="2024-01-31T14:01:56.486313Z",
+                    unread_count=1,
+                ),
                 "from": "27456897512",
                 "id": "ABCDEFGHIJKL_Ags-sF0gx5ts0DDMxw",
                 "text": {"body": "Hi there!"},
@@ -166,34 +189,16 @@ def voice_message():
         "contacts": [{"profile": {"name": "User"}, "wa_id": "27456897512"}],
         "messages": [
             {
-                "_vnd": {
-                    "v1": {
-                        "author": {"id": "27456897512", "name": "John Doe", "type": "OWNER"},
-                        "card_uuid": "None",
-                        "chat": {
-                            "assigned_to": "None",
-                            "contact_uuid": "ef59c32dsd9-289d-474e-aab1-e0dsss19f1f4",
-                            "inserted_at": "2024-01-25T09:02:46.684610Z",
-                            "owner": "+27456897512",
-                            "permalink": "https://whatsapp.turn.io/app/c/ef59c32dsd9-289d-474e-aab1-e0dsss19f1f4",
-                            "state": "OPEN",
-                            "state_reason": "Re-opened by inbound message.",
-                            "unread_count": 31,
-                            "updated_at": "2024-02-08T12:07:46.091528Z",
-                            "uuid": "08a64841-10df-4c11-b81f-4ec36d616c1c",
-                        },
-                        "direction": "inbound",
-                        "faq_uuid": "None",
-                        "in_reply_to": "None",
-                        "inserted_at": "2024-02-08T12:07:45.891699Z",
-                        "labels": [],
-                        "last_status": "None",
-                        "last_status_timestamp": "None",
-                        "on_fallback_channel": False,
-                        "rendered_content": "None",
-                        "uuid": "bd788d51-f3e1-11ff-e31d-0fc372a61d66",
-                    }
-                },
+                "_vnd": _vnd(
+                    author_name="John Doe",
+                    chat_uuid="08a64841-10df-4c11-b81f-4ec36d616c1c",
+                    contact_uuid="ef59c32dsd9-289d-474e-aab1-e0dsss19f1f4",
+                    permalink_uuid="ef59c32dsd9-289d-474e-aab1-e0dsss19f1f4",
+                    message_uuid="bd788d51-f3e1-11ff-e31d-0fc372a61d66",
+                    inserted_at="2024-02-08T12:07:45.891699Z",
+                    chat_updated_at="2024-02-08T12:07:46.091528Z",
+                    unread_count=31,
+                ),
                 "from": "27456897512",
                 "id": "ABGKLKLKLZd_Ags-DSDSdsWQUpsLqg",
                 "timestamp": "1707394065",
